Fixed rotor instructions, added rotor axes

+++ src/main/java/com/jverbruggen/jrides/animator/flatride/FlatRideComponentSpeed.java
@@ -1,11 +1,11 @@
 package com.jverbruggen.jrides.animator.flatride;
 
 public class FlatRideComponentSpeed {
-    private float speed;
-    private float minSpeed;
-    private float maxSpeed;
+    private double speed;
+    private double minSpeed;
+    private double maxSpeed;
 
-    public FlatRideComponentSpeed(float speed, float minSpeed, float maxSpeed) {
+    public FlatRideComponentSpeed(double speed, double minSpeed, double maxSpeed) {
         if(minSpeed > maxSpeed) throw new RuntimeException("Min speed has to be lower than max speed");
         else if(speed > maxSpeed) throw new RuntimeException("Speed has to be lower than max speed");
         else if(speed < minSpeed) throw new RuntimeException("Speed has to be higher than min speed");
@@ -15,7 +15,7 @@ public class FlatRideComponentSpeed {
         this.maxSpeed = maxSpeed;
     }
 
-    public FlatRideComponentSpeed(float speed){
+    public FlatRideComponentSpeed(double speed){
         this.speed = 0;
         this.minSpeed = 0;
         this.maxSpeed = speed;
@@ -25,17 +25,17 @@ public class FlatRideComponentSpeed {
      * Discouraged. This is an abrupt speed change.
      * Only use when absolutely necessary.
      */
-    public void setHard(float speed){
+    public void setHard(double speed){
         this.speed = speed;
     }
 
-    public void accelerate(float acceleration){
+    public void accelerate(double acceleration){
         this.speed += acceleration;
         if(this.speed > this.maxSpeed) this.speed = this.maxSpeed;
         else if(this.speed < this.minSpeed) this.speed = this.minSpeed;
     }
 
-    public void accelerateTowards(float acceleration, float towards){
+    public void accelerateTowards(double acceleration, double towards){
         if(this.speed > towards){
             this.speed -= acceleration;
             if(this.speed < towards) this.speed = towards;
@@ -48,15 +48,15 @@ public class FlatRideComponentSpeed {
         else if(this.speed < this.minSpeed) this.speed = this.minSpeed;
     }
 
-    public float getSpeed() {
+    public double getSpeed() {
         return speed;
     }
 
-    public float getMinSpeed() {
+    public double getMinSpeed() {
         return minSpeed;
     }
 
-    public float getMaxSpeed() {
+    public double getMaxSpeed() {
         return maxSpeed;
     }
